% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_hrv_nonlinear.R
\name{count_hrv_nonlinear}
\alias{count_hrv_nonlinear}
\title{Funtion to count hrv non-linear parameters sd1 and sd2 for examination object.}
\usage{
count_hrv_nonlinear(examination)
}
\arguments{
\item{examination}{Class object containig the information about the signal.
It contains two atributes:
*rr_signal - the vector of RR_intervals defined in miliseconds.
*hrv_time_domain - a list of heart rate variability parameters
                   (meanRR, minRR, maxRR, sdnn, rmssd).}
}
\value{
Poincare: a S3 class object containing two hrv nonlinear parameters: SD1 and SD2
}
\description{
Funtion to count hrv non-linear parameters sd1 and sd2 for examination object.
}
\examples{
examinationObj <- list(
  rr_signal = data.frame(signal = c(1000, 900, 1100, 950, 1050)),
  hrv_time_domain = list(meanRR = 1000, minRR = 900, maxRR = 1100, sdnn = 75.54)
)
class(examinationObj) <- "Examination"
hrv_Poincare <- count_hrv_nonlinear(examination = examinationObj)
}
