% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_hrv_time.R
\name{count_hrv_time}
\alias{count_hrv_time}
\title{Funtion to count hrv parameters in time domain for a given vector of
RR interval values.
Arguments:}
\usage{
count_hrv_time(rr_signal, unit = "miliseconds")
}
\arguments{
\item{rr_signal}{a dataframe containing a numerical vector of RR intervals
(time interval between two corresponding R beats in ECG signal).}

\item{unit}{a character value informing about the unit used in the rr_signal
vector. Can take two values: 'miliseconds' or 'seconds'.
The default value is set to 'miliseconds'.}
}
\value{
Examination: A S3 class object containig the information about the signal.
                      It contains two atributes:
                      *rr_signal - the vector of RR_intervals defined in miliseconds.
                      *hrv_time_domain - a list of heart rate variability parameters
                                         (meanRR, minRR, maxRR, sdnn, rmssd).
}
\description{
Funtion to count hrv parameters in time domain for a given vector of
RR interval values.
Arguments:
}
\examples{
data(rr01)
exam <- count_hrv_time(rr01, 'seconds')
}
