% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rr.R
\name{load_rr}
\alias{load_rr}
\title{Function to load rr signal either from local file or URL.}
\usage{
load_rr(path, header = FALSE)
}
\arguments{
\item{path}{A path to a local file or URL to a file from data source.}

\item{header}{A header of the file. Options:
FALSE - no header is loaded.
TRUE - first row of a file is assumed to be header
vector c('name1', 'name2'...) - given names of headers.
The default value is set to FALSE.}
}
\value{
rr_signal: list of rr intervals loaded from a file.
}
\description{
Function to load rr signal either from local file or URL.
}
\examples{
rr_signal <- load_rr(path='https://github.com/Mladog/rri/blob/main/rr01', header=c('RR'))
}
