% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr01.R
\docType{data}
\name{rr01}
\alias{rr01}
\title{RR data from https://physionet.org/content/challenge-2002}
\format{
A data frame with 1 column named 'RR' containing RR intervals [s].
}
\usage{
data(rr01)
}
\description{
The data set contains RR data.
}
\examples{
data(rr01)
plot(rr01$RR, type='line')
}
\keyword{RR}
\keyword{datasets}
\keyword{interval}
